DTLS 1.3 Update Epoch to a 64 bit counter and don't allow wrapping
This is based off a prior PR that was closed. The epoch was updated
for a 64 bit counter. A 16 bit epoch is kept when we read the
epoch from the wire in store it in the record layer.

Fixes: openssl/project#1796

+++ include/internal/recordmethod.h
@@ -116,7 +116,7 @@ struct ossl_record_method_st {
         const char *propq, int vers,
         int role, int direction,
         int level,
-        uint16_t epoch,
+        uint64_t epoch,
         unsigned char *secret,
         size_t secretlen,
         unsigned char *snkey,
@@ -324,7 +324,7 @@ struct ossl_record_method_st {
     /*
      * Get the Epoch value
      */
-    int (*get_epoch)(OSSL_RECORD_LAYER *rl, uint16_t *epoch);
+    int (*get_epoch)(OSSL_RECORD_LAYER *rl, uint64_t *epoch);
 
     /*
      * Set the current MTU length to be used for the record layer.
